# docs/swagger.yaml
basePath: /
definitions:
  controller.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      details: {}
      message:
        type: string
      traceId:
        type: string
    type: object
  dto.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  dto.LoginResponse:
    properties:
      accessToken:
        type: string
      expiresAt:
        type: string
    type: object
  dto.UserCreateRequest:
    properties:
      email:
        maxLength: 255
        type: string
      firstName:
        maxLength: 100
        type: string
      lastName:
        maxLength: 100
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
      role:
        maxLength: 50
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - firstName
    - lastName
    - password
    - role
    - username
    type: object
  dto.UserListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.UserResponse'
        type: array
      total:
        type: integer
    type: object
  dto.UserResponse:
    properties:
      createdAt:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      lastName:
        type: string
      role:
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
  dto.UserUpdateRequest:
    properties:
      email:
        maxLength: 255
        type: string
      firstName:
        maxLength: 100
        type: string
      lastName:
        maxLength: 100
        type: string
      role:
        maxLength: 50
        type: string
    required:
    - email
    - firstName
    - lastName
    - role
    type: object
info:
  contact: {}
  description: Simplified Go Web API boilerplate with JWT auth and SQL Server persistence.
  title: Go API Boilerplate
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return JWT
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.APIResponse'
      summary: Login
      tags:
      - Auth
  /api/v1/users:
    get:
      description: Paginated list of users
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserListResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.APIResponse'
      security:
      - BearerAuth: []
      summary: List users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/controller.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.APIResponse'
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - Users
  /api/v1/users/{id}:
    delete:
      description: Delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - Users
    get:
      description: Retrieve a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.APIResponse'
      security:
      - BearerAuth: []
      summary: Get user
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update existing user fields
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.APIResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - Users
  /healthz:
    get:
      description: Returns service health status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.APIResponse'
      summary: Health Check
      tags:
      - System
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"